standardizing file names
