+++ prototypes/_util_path.test.c
@@ -1,0 +1,26 @@
+#include "chepe-str.h"
+#include "env-utils.h"
+#include <stdio.h>
+#include <stdlib.h>
+
+/* function tester */
+int main(int argc, char **argv, char **env)
+{
+	(void) argc;
+	(void) argv;
+	(void) env;
+
+	linked_path *head = init_path_chain();
+	linked_path *node;
+
+	/* print_paths(); */
+
+	node = nav_path_chain(head);
+	do {
+		printf("%s\n", node->path);
+	} while ((node = nav_path_chain(NULL)));
+
+	erase_path_chain(head);
+
+	return (0);
+}
