+++ prototypes/util_env.c
@@ -1,5 +1,5 @@
-#include "chepe-str.h"
-#include "env-utils.h"
+#include "util_str.h"
+#include "util_env.h"
 #include <stdio.h>
 #include <stdlib.h>
