+++ prototypes/util_env.h
