+++ prototypes/util_str.h
